# expensetracker-apidocs.json
{"openapi":"3.0.1","info":{"title":"Expense Tracker API Documentation","description":"Backend APIs for Employee Expense & Revenue Tracking System","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/expenses/{id}":{"put":{"tags":["expense-controller"],"operationId":"updateExpense","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpenseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["expense-controller"],"operationId":"deleteExpense","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/expenses/approve/{id}":{"put":{"tags":["expense-controller"],"operationId":"approveExpense","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/register":{"post":{"tags":["user-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/employee":{"post":{"tags":["user-controller"],"operationId":"addEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/expenses":{"post":{"tags":["expense-controller"],"operationId":"addExpense","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpenseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/managers/{deptId}":{"get":{"tags":["user-controller"],"operationId":"getManagers","parameters":[{"name":"deptId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/employees":{"get":{"tags":["user-controller"],"operationId":"getEmployees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/expenses/user/{userId}":{"get":{"tags":["expense-controller"],"operationId":"getExpensesByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Expense"}}}}}}}},"/api/expenses/status/{status}":{"get":{"tags":["expense-controller"],"operationId":"getExpensesByStatus","parameters":[{"name":"status","in":"path","required":true,"schema":{"type":"string"}},{"name":"departmentId","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/departments":{"get":{"tags":["department-controller"],"operationId":"getDepartments","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Department"}}}}}}}},"/api/categories":{"get":{"tags":["category-controller"],"operationId":"getCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}}}}}}}},"components":{"schemas":{"ExpenseRequest":{"required":["amount","categoryId","date"],"type":"object","properties":{"userId":{"type":"integer","format":"int32"},"departmentId":{"type":"integer","format":"int32"},"categoryId":{"type":"integer","format":"int32"},"amount":{"type":"number","format":"double"},"date":{"type":"string"},"description":{"type":"string"},"managerId":{"type":"integer","format":"int32"}}},"ApproveRequest":{"required":["status"],"type":"object","properties":{"status":{"pattern":"(?i)approved|rejected|pending","type":"string"}}},"RegisterRequest":{"required":["email","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"},"role":{"type":"string"},"departmentId":{"type":"integer","format":"int32"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"Expense":{"type":"object","properties":{"expenseId":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"departmentId":{"type":"integer","format":"int32"},"categoryId":{"type":"integer","format":"int32"},"amount":{"type":"number","format":"double"},"date":{"type":"string"},"description":{"type":"string"},"receiptUrl":{"type":"string"},"status":{"type":"string"},"approvedBy":{"type":"integer","format":"int32"},"remarks":{"type":"string"}}},"Department":{"type":"object","properties":{"departmentId":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"Category":{"type":"object","properties":{"categoryId":{"type":"integer","format":"int32"},"name":{"type":"string"}}}}}}